--- sprint_1_webservice/src/main/java/api/controllers/SaleReportController.java
@@ -1,35 +1,0 @@
-package api.controllers;
-
-/*
-    Created by HauPV
-    Time: 19:16 31/05/2022
-    Function: display list sale report
-*/
-
-import api.models.ISaleReport;
-import api.services.ISaleReportService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@CrossOrigin("http://localhost:4200")
-@RequestMapping("/api/sale-report")
-public class SaleReportController {
-
-    @Autowired
-    ISaleReportService iSaleReportService;
-
-    @GetMapping("")
-    public ResponseEntity<List<ISaleReport>> saleReportList() {
-        List<ISaleReport> saleReportList = this.iSaleReportService.findAllSaleReports();
-        return new ResponseEntity<>(saleReportList, HttpStatus.OK);
-    }
-
-}
