+++ sprint_1_webservice/src/main/resources/application.properties
@@ -1,10 +1,6 @@
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/sprint-1-db?createDatabaseIfNotExist=true
 spring.datasource.username=root
-<<<<<<< HEAD
-spring.datasource.password=anhhieu
-=======
 spring.datasource.password=codegym
->>>>>>> eb36497fb9efc96985deb32a3c453fbd1b264623
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
