+++ sprint_1_webservice/src/test/java/api/controllers/SaleReportRestController_saleReportList.java
@@ -1,0 +1,30 @@
+package api.controllers;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.test.context.junit.jupiter.web.SpringJUnitWebConfig;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@AutoConfigureMockMvc
+public class SaleReportRestController_saleReportList {
+
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Test
+    public void saleReportList_5(){
+        try {
+            mockMvc.perform(MockMvcRequestBuilders.get("/api/sale-report")).andDo(print())
+                    .andExpect(status().is4xxClientError());
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+}
